# api/swagger-spec.json
{"openapi":"3.0.0","info":{"title":"iGrant POC","description":"iGrant API Documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"access-token":{"scheme":"bearer","bearerFormat":"Token","type":"http"}},"schemas":{"LoginRequestDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateGranteeMasterDto":{"type":"object","properties":{"grantee_name":{"type":"string"},"occupation":{"type":"string"},"name_of_institution":{"type":"string"},"ein":{"type":"string"},"address":{"type":"string"},"contact_details":{"type":"string"}},"required":["grantee_name","occupation","name_of_institution","ein","address","contact_details"]},"UpdateGranteeMasterDto":{"type":"object","properties":{"grantee_name":{"type":"string"},"occupation":{"type":"string"},"name_of_institution":{"type":"string"},"ein":{"type":"string"},"address":{"type":"string"},"contact_details":{"type":"string"}}},"CreateGranteeGrantDto":{"type":"object","properties":{}},"RenewGrantDTO":{"type":"object","properties":{}},"UpdateGranteeGrantDto":{"type":"object","properties":{}},"CreateGrantMasterDto":{"type":"object","properties":{"grant_name":{"type":"string"},"grantee_master_id":{"type":"object"},"grant_id":{"type":"string"},"grant_authorized":{"type":"number"},"agency":{"type":"string"},"program":{"type":"string"}},"required":["grant_name","grantee_master_id","grant_id","grant_authorized","agency","program"]},"UpdateGrantMasterDto":{"type":"object","properties":{"grant_name":{"type":"string"},"grantee_master_id":{"type":"object"},"grant_id":{"type":"string"},"grant_authorized":{"type":"number"},"agency":{"type":"string"},"program":{"type":"string"}}},"CreateAnswerDto":{"type":"object","properties":{"grantee_grant_id":{"type":"object"},"answer":{"type":"array","items":{"type":"string"}},"is_submit":{"type":"boolean"}},"required":["grantee_grant_id","answer","is_submit"]},"CreateQuestionMasterDto":{"type":"object","properties":{}},"UpdateQuestionMasterDto":{"type":"object","properties":{}},"CreateAuditDto":{"type":"object","properties":{}},"UpdateAuditDto":{"type":"object","properties":{}}}},"paths":{"/api/dashboard":{"get":{"operationId":"DashboardController_fetchDashboardDetails","parameters":[],"responses":{"200":{"description":"Fetch All Grant"},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]}},"/api/users/me":{"get":{"operationId":"UsersController_getCurrentUser","parameters":[],"responses":{"200":{"description":"A JSON Object containing user information"},"401":{"description":"No / Invalid access token was passed through the header"}},"security":[{"bearer":[]}]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDto"}}}},"responses":{"200":{"description":"User successfully logged in"},"400":{"description":"Bad Request Body"},"401":{"description":"Invalid Credentials"}}}},"/api/grantee-master/create-grantee":{"post":{"operationId":"GranteeMasterController_createGrantee","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGranteeMasterDto"}}}},"responses":{"200":{"description":"Creates New Grantee"},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]}},"/api/grantee-master":{"get":{"operationId":"GranteeMasterController_findAll","parameters":[],"responses":{"200":{"description":"Fetch All Grantee"},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]}},"/api/grantee-master/singleMultiGrant":{"get":{"operationId":"GranteeMasterController_singleMultiGrantFetch","parameters":[],"responses":{"200":{"description":"Fetch Single Or Multi Grant"},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]}},"/api/grantee-master/getAgencyGranteeList":{"get":{"operationId":"GranteeMasterController_findAllGranteeFOrAgency","parameters":[],"responses":{"200":{"description":"Fetch All Grantee"},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]}},"/api/grantee-master/{id}":{"get":{"operationId":"GranteeMasterController_findOne","parameters":[],"responses":{"200":{"description":"Fetch Single Grantee"},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]},"patch":{"operationId":"GranteeMasterController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGranteeMasterDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"GranteeMasterController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/grantee-grant":{"post":{"operationId":"GranteeGrantController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGranteeGrantDto"}}}},"responses":{"200":{"description":"Fetch Grant Details "},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]},"get":{"operationId":"GranteeGrantController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/api/grantee-grant/renew-grant":{"post":{"operationId":"GranteeGrantController_renewGrant","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RenewGrantDTO"}}}},"responses":{"200":{"description":"Fetch Grant Details "},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]}},"/api/grantee-grant/{id}":{"get":{"operationId":"GranteeGrantController_findGrantsForRespectiveGrant","parameters":[],"responses":{"200":{"description":"Fetch Grant Details "},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]},"patch":{"operationId":"GranteeGrantController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGranteeGrantDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"GranteeGrantController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/grantee-grant/risk_analysis/{id}":{"get":{"operationId":"GranteeGrantController_findRiskForGrant","parameters":[],"responses":{"200":{"description":"Fetch Grant Details "},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]}},"/api/grant-master/create-grant":{"post":{"operationId":"GrantMasterController_createGrant","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGrantMasterDto"}}}},"responses":{"200":{"description":"Creates New Grant "},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]}},"/api/grant-master/get-all-grant":{"get":{"operationId":"GrantMasterController_fetchAllGrant","parameters":[],"responses":{"200":{"description":"Fetch All Grant"},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]}},"/api/grant-master/{id}":{"patch":{"operationId":"GrantMasterController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGrantMasterDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"GrantMasterController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/answer":{"post":{"operationId":"AnswerController_createGrant","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAnswerDto"}}}},"responses":{"200":{"description":"Push answers submitted by Grantors"},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]}},"/api/answer/fetch-file":{"get":{"operationId":"AnswerController_getFileAsBuffer","parameters":[],"responses":{"200":{"description":"Push answers submitted by Grantors"},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]}},"/api/question-master":{"post":{"operationId":"QuestionMasterController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateQuestionMasterDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"QuestionMasterController_fetchAll","parameters":[],"responses":{"200":{"description":"Creates New Grantee"},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]}},"/api/question-master/{id}":{"get":{"operationId":"QuestionMasterController_fetchQuestions","parameters":[],"responses":{"200":{"description":"Creates New Grantee"},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]},"patch":{"operationId":"QuestionMasterController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateQuestionMasterDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"QuestionMasterController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/question-master/questionAdd":{"post":{"operationId":"QuestionMasterController_postSno","parameters":[],"responses":{"201":{"description":""}}}},"/api/audit":{"post":{"operationId":"AuditController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAuditDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"AuditController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/api/audit/{id}":{"get":{"operationId":"AuditController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"AuditController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAuditDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"AuditController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/audit/push-dummy":{"post":{"operationId":"AuditController_pushDummy","parameters":[],"responses":{"200":{"description":""}}}},"/api/audit/audit-upload":{"post":{"operationId":"AuditController_manualUpload","parameters":[],"responses":{"200":{"description":"Upload Audit File"},"401":{"description":"No / Invalid access token was passed through the header"},"403":{"description":"User's role doesn't enable them to access this feature"}},"security":[{"bearer":[]}]}}}}